simply edits component ahead of refactor
