+++ front/app/components/Edits/index.ts
@@ -1,1 +1,2 @@
-export { default } from './Edits';
+export { default } from "./Edits";
+export { default as WikiPageEditFragment } from "./WikiPageEditFragment";
