+++ front/app/containers/WikiPage/WikiPage.tsx
@@ -16,7 +16,7 @@ import * as FontAwesome from 'react-fontawesome';
 
 import LoadingPane from 'components/LoadingPane';
 import Error from 'components/Error';
-import Edits from 'components/Edits';
+import Edits, { WikiPageEditFragment } from 'components/Edits';
 import { trimPath } from 'utils/helpers';
 import CurrentUser from 'containers/CurrentUser';
 import { UserFragment } from 'types/UserFragment';
@@ -49,7 +49,7 @@ const FRAGMENT = gql`
     meta
   }
 
-  ${Edits.fragment}
+  ${WikiPageEditFragment}
 `;
 
 const QUERY = gql`
@@ -88,11 +88,11 @@ const Toolbar = styled.div`
   padding: 10px;
 `;
 
-class QueryComponent extends Query<WikiPageQuery, WikiPageQueryVariables> {}
+class QueryComponent extends Query<WikiPageQuery, WikiPageQueryVariables> { }
 class UpdateContentMutation extends Mutation<
   WikiPageUpdateContentMutation,
   WikiPageUpdateContentMutationVariables
-> {}
+  > { }
 
 class WikiPage extends React.Component<WikiPageProps, WikiPageState> {
   state: WikiPageState = {
@@ -118,20 +118,20 @@ class WikiPage extends React.Component<WikiPageProps, WikiPageState> {
     this.props.onLoaded && this.props.onLoaded();
   };
 
-  handleHistory = (hash:string, siteViewUrl:string) => {
+  handleHistory = (hash: string, siteViewUrl: string) => {
     this.props.history.push(`${trimPath(this.props.match.url)}/wiki/history?hash=${hash}&sv=${siteViewUrl}`);
   };
 
-  handleEdit = (hash:string, siteViewUrl:string) => {
-    console.log("this.props",this.props.match)
+  handleEdit = (hash: string, siteViewUrl: string) => {
+    console.log("this.props", this.props.match)
     this.props.history.push(`${trimPath(this.props.match.url)}/wiki/edit?hash=${hash}&sv=${siteViewUrl}`);
   };
 
-  handleView = (hash:string, siteViewUrl:string) => {
+  handleView = (hash: string, siteViewUrl: string) => {
     this.props.history.push(`${trimPath(this.props.match.url)}?hash=${hash}&sv=${siteViewUrl}`);
   };
 
-  handlePreview = (hash:string, siteViewUrl:string) => {
+  handlePreview = (hash: string, siteViewUrl: string) => {
     if (this.state.editorState === 'plain') {
       const text = this.getEditorText() || '';
       this.setState({
@@ -199,13 +199,13 @@ class WikiPage extends React.Component<WikiPageProps, WikiPageState> {
     );
   };
 
-  renderEditButton = (isAuthenticated: boolean, hash:string, siteViewUrl:string) => {
+  renderEditButton = (isAuthenticated: boolean, hash: string, siteViewUrl: string) => {
     if (!isAuthenticated) return null;
 
     return (
       <Button
         type="button"
-        onClick={()=>this.handleEdit(hash, siteViewUrl)}
+        onClick={() => this.handleEdit(hash, siteViewUrl)}
         style={{ marginLeft: '10px' }}>
         Edit <FontAwesome name="edit" />
       </Button>
@@ -232,7 +232,7 @@ class WikiPage extends React.Component<WikiPageProps, WikiPageState> {
     );
   };
 
-  renderToolbar = (data: WikiPageQuery, user: UserFragment | null, hash:string, siteViewUrl:string) => {
+  renderToolbar = (data: WikiPageQuery, user: UserFragment | null, hash: string, siteViewUrl: string) => {
     const isAuthenticated = user !== null;
     return (
       <Toolbar>
@@ -244,7 +244,7 @@ class WikiPage extends React.Component<WikiPageProps, WikiPageState> {
                 {this.renderMarkdownButton()}{' '}
                 <Button
                   type="button"
-                  onClick={()=>this.handlePreview(hash, siteViewUrl)}
+                  onClick={() => this.handlePreview(hash, siteViewUrl)}
                   style={{ marginLeft: '10px' }}>
                   Preview <FontAwesome name="photo" />
                 </Button>
@@ -259,7 +259,7 @@ class WikiPage extends React.Component<WikiPageProps, WikiPageState> {
                 {this.renderEditButton(isAuthenticated, hash, siteViewUrl)}{' '}
                 <Button
                   type="button"
-                  onClick={()=>this.handleView(hash,siteViewUrl)}
+                  onClick={() => this.handleView(hash, siteViewUrl)}
                   style={{ marginLeft: '10px' }}>
                   View <FontAwesome name="photo" />
                 </Button>
@@ -270,7 +270,7 @@ class WikiPage extends React.Component<WikiPageProps, WikiPageState> {
           <Route
             render={() => (
               <>
-                <Button type="button" onClick={()=>this.handleHistory(hash, siteViewUrl)}>
+                <Button type="button" onClick={() => this.handleHistory(hash, siteViewUrl)}>
                   History <FontAwesome name="history" />
                 </Button>
                 {this.renderEditButton(isAuthenticated, hash, siteViewUrl)}
